back-end/requirements.txt
absl-py==1.4.0
asgiref==3.6.0
boto3==1.26.156
botocore==1.29.156
cachetools==5.3.1
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.10.0
distlib==0.3.6
Django==3.2.19
django-cors-headers==4.1.0
djangorestframework==3.14.0
filelock==3.12.0
fire==0.5.0
fonttools==4.39.3
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.20.0
google-auth-oauthlib==1.0.0
grpcio==1.54.2
huggingface-hub==0.15.1
idna==2.10
importlib-metadata==6.7.0
importlib-resources==5.12.0
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
mssql-django==1.2
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
platformdirs==3.5.1
protobuf==4.23.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybboxes==0.1.6
pyodbc==4.0.39
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.29.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
roboflow==1.0.9
rsa==4.9
s3transfer==0.6.1
sahi==0.11.14
scipy==1.10.1
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
smmap==5.0.0
sqlparse==0.4.4
sympy==1.11.1
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
ultralytics==8.0.120
urllib3==1.26.15
virtualenv==20.23.0
Werkzeug==2.3.6
wget==3.2
yolov5==7.0.12
zipp==3.15.0
